clean up nlp folder
